requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
